Move repository() to audb.core.api
