+++ audb/core/api.py
@@ -13,9 +13,11 @@ from audb.core import define
 from audb.core.config import config
 from audb.core.dependencies import Dependencies
 from audb.core.flavor import Flavor
+from audb.core.repository import Repository
 from audb.core.utils import (
     lookup_backend,
     mix_mapping,
+    _lookup,
 )
 
 
@@ -574,6 +576,35 @@ def remove_media(
                     remote_archive,
                     version,
                 )
+
+
+def repository(
+        name: str,
+        version: str,
+) -> Repository:
+    r"""Return repository that stores the requested database.
+
+    If the database is stored in several repositories,
+    only the first one is returned.
+    The order of the repositories to look for the database
+    is given by :attr:`config.REPOSITORIES`.
+
+    Args:
+        name: database name
+        version: version string
+
+    Returns:
+        repository that contains the database
+
+    Raises:
+        RuntimeError: if database is not found
+
+    Example:
+        >>> audb.repository('emodb', '1.1.1')
+        Repository('data-public', 'https://audeering.jfrog.io/artifactory', 'artifactory')
+
+    """  # noqa: E501
+    return _lookup(name, version)[0]
 
 
 def versions(
